+++ sys/Graphics_record.cpp
@@ -1,6 +1,6 @@
 /* Graphics_record.cpp
  *
- * Copyright (C) 1992-2011,2013,2014,2015,2016,2017 Paul Boersma
+ * Copyright (C) 1992-2005,2007-2019 Paul Boersma
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -195,7 +195,7 @@ void Graphics_play (Graphics me, Graphics thee) {
 				Graphics_setFont (thee, (enum kGraphics_font) get);
 			} break;
 			case SET_FONT_SIZE: {
-				Graphics_setFontSize (thee, (int) get);
+				Graphics_setFontSize (thee, get);
 			} break;
 			case SET_FONT_STYLE: {
 				Graphics_setFontStyle (thee, (int) get);
